fix(windows): harden WebView2 staleness guard, wakelock scope, C++ narrowing
- Reader onReceivedError: mirror onLoadStop's chapter snapshot, lyrics
  mode check and URL path comparison so stale callbacks are dropped
- WakelockPlus.disable() at startup: remove mobile-only guard since
  the plugin supports Windows (SetThreadExecutionState)
- Dictionary popup: log a debugPrint when Windows asset inlining fails
  and the fallback file:// path is taken
- hoshidicts C++: add explicit static_cast on all write_val<T>() calls
  that narrow size_t, plus bounds checks for tags/rules/media path to
  prevent silent truncation — eliminates all MSVC C4267 warnings

+++ native/hoshidicts/hoshidicts_src/stardict/stardict_reader.cpp
@@ -192,7 +192,7 @@ StardictResult stardict_reader::parse_from_data(
             while (dpos + len < entry_size && dict_data[offset + dpos + len] != 0) len++;
             if (!definition.empty()) definition += '\n';
             definition.append(start, len);
-            dpos += (dpos + len < entry_size) ? len + 1 : len;
+            dpos += static_cast<uint32_t>((dpos + len < entry_size) ? len + 1 : len);
           }
         } else {
           if (last_field) {
